selection shortcuts + misc bugs fixed

+++ Assets/Scripts/Entitas/Components/Game/Attack/WeaponComponent.cs
@@ -6,5 +6,5 @@ public class WeaponComponent : IComponent
     public int range;
     public float rotationSpeed;
     public float rateOfFire;
-    public float dmg;
+    public int dmg;
 }
